+++ src/main/scala/com/github/oen9/slinky/bridge/konva/Konva.scala
@@ -1,7 +1,8 @@
 package com.github.oen9.slinky.bridge.konva
 
-import scalajs.js.annotation.JSImport
+import com.github.oen9.slinky.bridge.reactkonva.Operations
 import scalajs.js
+import scalajs.js.annotation.JSImport
 
 @JSImport("konva", JSImport.Default)
 @js.native
@@ -16,13 +17,28 @@ object Konva extends js.Object {
     }
     @js.native
     trait ClientRect extends js.Object {
-      def x: Int
-      def y: Int
-      def width: Int
-      def height: Int
+      def x: Double
+      def y: Double
+      def width: Double
+      def height: Double
     }
     def getRandomColor(): String = js.native
     def getRGB(color: String): RGB = js.native
     def haveIntersection(r1: ClientRect, r2: ClientRect): Boolean = js.native
+  }
+
+  @js.native
+  class Animation(func: js.Function1[Frame, Unit], layers: Seq[Operations.Ref]) extends js.Object {
+    def start(): Unit = js.native
+    def stop(): Unit = js.native
+    def isRunning(): Boolean = js.native
+  }
+
+  @js.native
+  trait Frame extends js.Object {
+    def timeDiff: Double
+    def lastTime: Double
+    def time: Double
+    def frameRate : Double
   }
 }