+++ src/main/scala/com/github/oen9/slinky/bridge/reactkonva/Layer.scala
@@ -1,7 +1,7 @@
 package com.github.oen9.slinky.bridge.reactkonva
 
-import slinky.core.ExternalComponentNoProps
+import slinky.core.ExternalComponentNoPropsWithRefType
 
-object Layer extends ExternalComponentNoProps {
+object Layer extends ExternalComponentNoPropsWithRefType[Operations.Ref] {
   override val component = ReactKonvaDOM.Layer
 }